Use u32 for metadata.payload_size
As, on 32-bit architectures, u64 doesn't satisfy `ToUsize` that `nom` expects.

+++ src/message.rs
@@ -497,7 +497,7 @@ fn batched_message(i: &[u8]) -> IResult<&[u8], BatchedMessage> {
         |metadata| metadata.payload_size >= 0,
     )(i)?;
 
-    let (i, payload) = take(metadata.payload_size as u64)(i)?;
+    let (i, payload) = take(metadata.payload_size as u32)(i)?;
 
     Ok((
         i,
